[Qt][WK2] Caching of ShareableSurfaces breaks tiling. https://bugs.webkit.org/show_bug.cgi?id=91609
A ShareableSurface should only be cached,
when it is GraphicsSurface based.

Reviewed by Kenneth Rohde Christiansen.

* UIProcess/LayerTreeCoordinatorProxy.cpp:
(WebKit::LayerTreeCoordinatorProxy::updateTileForLayer):
* UIProcess/LayerTreeCoordinatorProxy.h:
(LayerTreeCoordinatorProxy):

Canonical link: https://commits.webkit.org/109406@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@122948 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebKit2/UIProcess/LayerTreeCoordinatorProxy.h
@@ -79,7 +79,7 @@ protected:
     DrawingAreaProxy* m_drawingAreaProxy;
     RefPtr<WebLayerTreeRenderer> m_renderer;
 #if USE(GRAPHICS_SURFACE)
-    HashMap<uint64_t, RefPtr<ShareableSurface> > m_surfaces;
+    HashMap<uint32_t, RefPtr<ShareableSurface> > m_surfaces;
 #endif
 };
